requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
backports-zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
oauthlib==3.2.2
openai==1.14.2
pip==24.0
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pyjwt==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.2.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
youtube-transcript-api==0.6.2
python-dotenv=1.0.1